+++ src/main/java/kitchenpos/menu/domain/MenuValidator.java
@@ -8,6 +8,8 @@ import org.springframework.util.CollectionUtils;
 
 import kitchenpos.menu.dto.MenuProductRequest;
 import kitchenpos.menu.dto.MenuRequest;
+import kitchenpos.product.domain.Product;
+import kitchenpos.product.domain.ProductRepository;
 
 @Component
 public class MenuValidator {
