+++ src/main/java/kitchenpos/product/dto/ProductRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.dto;
+package kitchenpos.product.dto;
 
 import java.math.BigDecimal;
