+++ src/main/java/kitchenpos/product/ui/ProductRestController.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.ui;
+package kitchenpos.product.ui;
 
 import java.net.URI;
 import java.util.List;
@@ -12,9 +12,9 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import kitchenpos.menu.application.ProductService;
-import kitchenpos.menu.dto.ProductRequest;
-import kitchenpos.menu.dto.ProductResponse;
+import kitchenpos.product.application.ProductService;
+import kitchenpos.product.dto.ProductRequest;
+import kitchenpos.product.dto.ProductResponse;
 
 @RestController
 @RequestMapping("/api/products")
